auth: Refactor auth files Description: Extracted auth related code and transferred to `shield`
Expected behavior:
able to implement the shield db in shield middleware
