+++ src/main/java/com/splitscale/fordastore/core/repositories/UserRepository.java
@@ -2,7 +2,6 @@ package com.splitscale.fordastore.core.repositories;
 
 import com.splitscale.fordastore.core.user.User;
 import com.splitscale.fordastore.core.user.UserRequest;
-import com.splitscale.fordastore.core.user.register.UserClaims;
 
 public interface UserRepository {
 
@@ -10,7 +9,7 @@ public interface UserRepository {
 
   public User findByUsername(String username) throws Exception;
 
-  public UserClaims add(UserRequest userRequest) throws Exception;
+  public User add(UserRequest userRequest) throws Exception;
 
   public boolean update(Long id, UserRequest userRequest) throws Exception;
 }
