--- src/main/java/com/splitscale/fordastore/core/security/Authorization.java
@@ -1,9 +1,0 @@
-package com.splitscale.fordastore.core.security;
-
-import java.security.PublicKey;
-
-public interface Authorization {
-  PublicKey getPublicKey();
-
-  String getToken();
-}
