--- src/main/java/com/splitscale/fordastore/core/security/EncryptionService.java
@@ -1,9 +1,0 @@
-package com.splitscale.fordastore.core.security;
-
-public interface EncryptionService {
-
-  String encrypt(String input);
-
-  boolean verify(String password, String hash);
-
-}
